filter and adding proofType table

+++ HotelApp.Application/Services/RoleService/RoleService.cs
@@ -11,25 +11,29 @@ using HotelApp.Domain.Entities;
 
 namespace HotelApp.Application.Services.RoleService
 {
-    public class RoleService : IRoleService
+	public class RoleService : IRoleService
 	{
 		private readonly RoleManager<Role> _roleManager;
 		private readonly IMapper _mapper;
 		private readonly IRoleRepository _roleRepository;
 		private readonly UserManager<User> _userManager;
-		private readonly int _userId;
+		private readonly IUnitOfWork _unitOfWork;
+		private readonly IHttpContextAccessor _httpContextAccessor;
 
 		public RoleService(
 			RoleManager<Role> roleManager,
 			IMapper mapper,
 			IRoleRepository roleRepository,
-			UserManager<User> userManager)
+			UserManager<User> userManager,
+			IUnitOfWork unitOfWork,
+			IHttpContextAccessor httpContextAccessor)
 		{
 			_roleManager = roleManager;
 			_mapper = mapper;
 			_roleRepository = roleRepository;
 			_userManager = userManager;
-			_userId = 1; 
+			_unitOfWork = unitOfWork;
+			_httpContextAccessor = httpContextAccessor;
 		}
 		// ✅ Get all available permissions with info about which are assigned
 		public async Task<IEnumerable<PermissionDTO>> GetAllPermissionsAsync(int roleId)
@@ -46,8 +50,17 @@ namespace HotelApp.Application.Services.RoleService
 		// ✅ Get all roles
 		public async Task<IEnumerable<GetRolesDTO>> GetRolesAsync()
 		{
-			var roles = await _roleManager.Roles.ToListAsync();
-			return _mapper.Map<List<GetRolesDTO>>(roles);
+			var roles = await _roleManager.Roles.Select(r => new GetRolesDTO
+			{
+				Id = r.Id.ToString(),
+				Name = r.Name,
+				CreatedDate = r.CreatedDate,
+				LastModifiedDate = r.LastModifiedDate,
+				IsActive = r.IsActive,
+				IsBasic = r.IsBasic
+			})
+			.ToListAsync();
+			return roles;
 		}
 
 		// ✅ Get a single role
@@ -68,7 +81,7 @@ namespace HotelApp.Application.Services.RoleService
 			{
 				Name = roleDTO.Name,
 				IsActive = roleDTO.IsActive,
-				CreatedById = _userId,
+				CreatedById = GetCurrentUserId(),
 				CreatedDate = DateTime.UtcNow
 			};
 
@@ -82,24 +95,6 @@ namespace HotelApp.Application.Services.RoleService
 			return ServiceResponse<RoleDTO>.ResponseSuccess($"Role '{roleDTO.Name}' created successfully");
 		}
 
-		
-		//// ✅ Assign selected permissions to a role
-		//public async Task<ServiceResponse<string>> AssignPermissionsAsync(int roleId, List<string> permissions)
-		//{
-		//	var role = await GetRoleByIdAsync(roleId);
-		//	if (role == null)
-		//		return ServiceResponse<string>.ResponseFailure($"Role ID {roleId} not found");
-
-		//	await _roleRepository.RemoveAllRolePermissionsAsync(role.Id);
-		//	await _roleRepository.AddRolePermissionsAsync(role.Id, permissions);
-
-		//	role.LastModifiedById = _userId;
-		//	role.LastModifiedDate = DateTime.UtcNow;
-		//	await _roleManager.UpdateAsync(role);
-
-		//	return ServiceResponse<string>.ResponseSuccess($"Role '{role.Name}' permissions updated successfully");
-		//}
-		// ✅ Get detailed permissions grouped from JSON and assigned status
 		public async Task<AssignPermissionDTO> GetAssignPermissionsDTOAsync(int roleId)
 		{
 			var role = await GetRoleByIdAsync(roleId);
@@ -126,15 +121,34 @@ namespace HotelApp.Application.Services.RoleService
 			};
 		}
 
-		// ✅ Assign permissions to role (replace old with new)
 		public async Task<ServiceResponse<object>> AssignPermissionsAsync(int roleId, List<string> newPermissions)
 		{
+            bool IsExist = await _unitOfWork.Repository<Role>()
+                .IsExistsAsync(r => r.Id == roleId);
+            if (!IsExist)
+            {
+                return ServiceResponse<object>.ResponseFailure("Role not exist!");
+            }
+
+            bool IsBasic = await _unitOfWork.Repository<Role>()
+				.IsExistsAsync(r => r.Id == roleId && r.IsBasic);
+			if (IsBasic)
+			{
+				return ServiceResponse<object>.ResponseFailure("Can\'t update basic Role.");
+			}
+
 			try
 			{
 				await _roleRepository.RemoveAllRolePermissionsAsync(roleId);
 				await _roleRepository.AddRolePermissionsAsync(roleId, newPermissions);
 
-				return ServiceResponse<object>.ResponseSuccess("Permissions Updated Successfully.");
+                var role = await _roleManager.FindByIdAsync(roleId.ToString());
+                role.LastModifiedById = GetCurrentUserId();
+				role.LastModifiedDate = DateTime.UtcNow;
+
+				await _roleManager.UpdateAsync(role);
+
+                return ServiceResponse<object>.ResponseSuccess("Permissions Updated Successfully.");
 			}
 			catch (Exception ex)
 			{
@@ -142,7 +156,7 @@ namespace HotelApp.Application.Services.RoleService
 			}
 		}
 
-		// ✅ Edit role name / activation
+		// ✅ Edit role 
 		public async Task<ServiceResponse<RoleDTO>> EditRoleAsync(Role role)
 		{
 			var existingRole = await _roleManager.FindByIdAsync(role.Id.ToString());
@@ -154,7 +168,7 @@ namespace HotelApp.Application.Services.RoleService
 
 			existingRole.Name = role.Name;
 			existingRole.IsActive = role.IsActive;
-			existingRole.LastModifiedById = _userId;
+			existingRole.LastModifiedById = GetCurrentUserId();
 			existingRole.LastModifiedDate = DateTime.UtcNow;
 
 			var result = await _roleManager.UpdateAsync(existingRole);
@@ -164,7 +178,7 @@ namespace HotelApp.Application.Services.RoleService
 			return ServiceResponse<RoleDTO>.ResponseSuccess("Role updated successfully");
 		}
 
-		// ✅ Delete role (if safe)
+		// ✅ Delete role
 		public async Task<ServiceResponse<Role>> DeleteRoleAsync(int roleId)
 		{
 			var role = await _roleManager.FindByIdAsync(roleId.ToString());
@@ -186,197 +200,17 @@ namespace HotelApp.Application.Services.RoleService
 		}
 
 
+
+		#region Helper Methods 
+		private int GetCurrentUserId()
+		{
+			if (int.TryParse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier), out int userId))
+			{
+				return userId;
+			}
+
+			return 0;
+		}
+		#endregion
 	}
-
-
-
-
-
-
-	//   public class RoleService : IRoleService
-	//{
-	//	private readonly RoleManager<Role> _roleManager;
-	//	private readonly IMapper _mapper;
-	//	private readonly IRoleRepository _roleRepository;
-	//	private readonly IUnitOfWork _unitOfWork;
-	//	private readonly UserManager<User> _userManager;
-
-	//       private readonly IHttpContextAccessor _httpContextAccessor;
-	//       private readonly int? _userId;
-	//       public RoleService(RoleManager<Role> roleManager, IMapper mapper,
-	//		IRoleRepository roleRepository, IUnitOfWork unitOfWork, UserManager<User> userManager,
-	//           IHttpContextAccessor httpContextAccessor)
-	//	{
-	//		_roleManager = roleManager;
-	//		_mapper = mapper;
-	//		_roleRepository = roleRepository;
-	//		_unitOfWork = unitOfWork;
-	//		_userManager = userManager;
-	//		_httpContextAccessor = httpContextAccessor;
-
-	//           //_userId = int.Parse(_httpContextAccessor.HttpContext?.User?.FindFirst(ClaimTypes.NameIdentifier)?.Value);
-	//       }
-
-
-	//       public async Task<IEnumerable<GetRolesDTO>> GetRolesAsync()
-	//	{
-	//		var roles = await _unitOfWork.Repository<Role>().GetAllAsDtoAsync<GetRolesDTO>();
-
-	//		return roles;
-	//	}
-
-	//	public async Task<Role?> GetRoleByIdAsync(int roleId)
-	//	{
-	//		return await _roleManager.FindByIdAsync(roleId.ToString());
-	//	}
-
-	//	public async Task<ServiceResponse<RoleDTO>> AddRoleAsync(RoleDTO roleDTO)
-	//	{
-	//		try
-	//		{
-	//			if (await _roleManager.RoleExistsAsync(roleDTO.Name) == true)
-	//			{
-	//				return ServiceResponse<RoleDTO>.ResponseFailure($"The role '{roleDTO.Name}' is alresdy exists");
-	//			}
-	//			Role newRole = new Role()
-	//			{
-	//				Name = roleDTO.Name,
-	//				IsActive = roleDTO.IsActive,
-	//				CreatedById = _userId,
-	//				CreatedDate = DateTime.UtcNow
-	//			};
-
-	//			var result = await _roleManager.CreateAsync(newRole);
-	//			if (!result.Succeeded)
-	//			{
-	//				var roleErrors = result.Errors.Select(r => r.Description).ToList();
-	//				return ServiceResponse<RoleDTO>.ResponseFailure(string.Join(", ", roleErrors));
-	//			}
-	//			return ServiceResponse<RoleDTO>.ResponseSuccess($"Role '{roleDTO.Name}' created successfully");
-
-	//		}
-	//		catch (Exception ex)
-	//		{
-	//			return ServiceResponse<RoleDTO>.ResponseFailure($"An exception occurred: {ex.InnerException.Message}");
-
-	//		}
-	//	}
-
-	//	public async Task<IEnumerable<Entity>> GetAllEntitiesAsync()
-	//	{
-	//		return await _roleRepository.GetAllEntitiesAsync();
-	//	}
-
-	//	public async Task<IEnumerable<PermissionDTO>> GetAllPermissionsAsync(int roleId)
-	//	{
-	//		var assignedIds = await _roleRepository.GetAssignedPermissionsAsync(roleId);
-	//		var allPermissions = await _roleRepository.GetAllPermissionsAsync();
-
-	//		return allPermissions.Select(p => new PermissionDTO
-	//		{
-	//			Id = p.Id,
-	//			Action = p.Action,
-	//			EntityId = p.EntityId,
-	//			EntityName = p.Entity?.Name,
-	//			IsAssigned = assignedIds.Contains(p.Id)
-	//		});
-	//	}
-
-	//	public async Task<ServiceResponse<string>> AssignPermissionsAsync(int roleId, List<int> permissionIds)
-	//	{
-	//		var role = await GetRoleByIdAsync(roleId);
-	//		if (role == null)
-	//		{
-	//			return ServiceResponse<string>.ResponseFailure($"Role with ID {roleId} not found");
-	//		}
-	//		await _roleRepository.RemoveAllRolePermissionsAsync(roleId);
-	//		try
-	//		{
-	//			var permissionsToRemove = await _unitOfWork.Repository<RolePermission>().GetAllAsync(rp => rp.RoleId == roleId);
-
-	//			_unitOfWork.Repository<RolePermission>().DeleteRange(permissionsToRemove);
-	//			await _unitOfWork.CommitAsync();
-
-	//			var PermissionstoAdd = permissionIds.Select(permissionId => new RolePermission
-	//			{
-	//				RoleId = roleId,
-	//				PermissionId = permissionId
-	//               }).ToList();
-	//			await _unitOfWork.Repository<RolePermission>().AddRangeAsync(PermissionstoAdd);
-
-	//			await _unitOfWork.CommitAsync();
-
-	//			var ExistRole = await _roleManager.FindByIdAsync(roleId.ToString());
-	//               ExistRole.LastModifiedById = _userId;
-	//               ExistRole.LastModifiedDate = DateTime.UtcNow;
-
-	//			await _roleManager.UpdateAsync(ExistRole);
-
-	//			return ServiceResponse<string>.ResponseSuccess($"Role: {role.Name} Updated Succeffully");
-	//		}
-	//		catch (Exception ex)
-	//		{
-	//			return ServiceResponse<string>.ResponseFailure(ex.Message);
-	//		}
-	//	}
-
-
-
-	//	public async Task<ServiceResponse<RoleDTO>> EditRoleAsync(Role role)
-	//	{
-	//		var Oldrole = await _roleManager.FindByIdAsync(role.Id.ToString());
-	//		if (Oldrole == null)
-	//		{
-	//			return ServiceResponse<RoleDTO>.ResponseFailure($"Role con't be found");
-	//		}
-	//		if (Oldrole.IsBasic == true)
-	//		{
-	//			return ServiceResponse<RoleDTO>.ResponseFailure($"{role.Name} can't be Edited via permission");
-	//		}
-
-	//		Oldrole.Name = role.Name;
-	//		Oldrole.LastModifiedById = _userId;
-	//		Oldrole.LastModifiedDate = DateTime.UtcNow;
-	//		Oldrole.IsActive = role.IsActive;
-	//		try
-	//		{
-	//			var result = await _roleManager.UpdateAsync(Oldrole);
-	//			if (result.Succeeded)
-	//			{
-	//				return ServiceResponse<RoleDTO>.ResponseSuccess($"The Role is updated successfully");
-	//			}
-	//			return ServiceResponse<RoleDTO>.ResponseFailure(result.Errors.FirstOrDefault()?.Description ?? "Somthing went wrong!");
-	//		}
-	//		catch (Exception ex)
-	//		{
-	//			return ServiceResponse<RoleDTO>.ResponseFailure($"Error Occurred while proccessing: {ex.Message}");
-	//		}
-	//	}
-	//	public async Task<ServiceResponse<Role>> DeleteRoleAsync(int Id)
-	//	{
-	//		var role = await _roleManager.FindByIdAsync(Id.ToString());
-	//		if(role == null)
-	//		{
-	//			return ServiceResponse<Role>.ResponseFailure($"Role ID:{Id} con't be found");
-	//		}
-	//		if(role.IsBasic == true)
-	//		{
-	//			return ServiceResponse<Role>.ResponseFailure($"{role.Name} can't be deleted via permission");
-	//		}
-
-	//		var usersInRole = await _userManager.GetUsersInRoleAsync(role.Name);
-	//		if (usersInRole.Any())
-	//		{
-	//			return ServiceResponse<Role>.ResponseFailure($"The role '{role.Name}' cannot be deleted because it is assigned to one or more users.");
-	//		}
-
-	//		var result = await _roleManager.DeleteAsync(role);
-	//		if (result.Succeeded)
-	//		{
-	//			return ServiceResponse<Role>.ResponseSuccess($"{role.Name} Role deleted succeefully");
-	//		}
-
-	//		return ServiceResponse<Role>.ResponseFailure("Somthing is wrong with role deletion"); 
-	//	}
-	//}
 }
